# backend-api/internal/controller/user.go
package controller

import (
	"database/sql"

	"github.com/gin-gonic/gin"
	"github.com/tnnevol/openlist-strm/backend-api/internal/logger"
	"github.com/tnnevol/openlist-strm/backend-api/internal/middleware"
	"github.com/tnnevol/openlist-strm/backend-api/internal/service"
	"github.com/tnnevol/openlist-strm/backend-api/internal/util"
)

// SendCode godoc
// @Summary      发送验证码
// @Description  发送邮箱验证码
// @Tags         用户
// @Accept       json
// @Produce      json
// @Param        email  body  object{email=string}  true  "邮箱"
// @Success      200    {object}  model.Response
// @Router       /user/send-code [post]
func SendCode(db *sql.DB) gin.HandlerFunc {
	return func(c *gin.Context) {
		logger.Info("[API] /user/send-code called")
		var req struct {
			Email string `json:"email" binding:"required"`
		}
		if err := c.ShouldBindJSON(&req); err != nil {
			middleware.ValidationError(c, "邮箱不能为空")
			return
		}
		if !util.IsValidEmail(req.Email) {
			middleware.ValidationError(c, "邮箱格式不正确")
			return
		}
		code, err := service.SendCode(db, req.Email)
		if err != nil {
			if err.Error() == "用户已注册" {
				middleware.BadRequest(c, "用户已注册")
			} else {
				middleware.InternalServerError(c, "发送验证码失败")
			}
			return
		}
		middleware.SuccessWithMessage(c, "验证码已发送", gin.H{
			"code": code,
		})
	}
}

// Register godoc
// @Summary      注册并激活
// @Description  用户注册并激活账户
// @Tags         用户
// @Accept       json
// @Produce      json
// @Param        register body object{email=string,password=string,code=string} true "注册信息"
// @Success      200    {object}  model.Response
// @Router       /user/register [post]
func Register(db *sql.DB) gin.HandlerFunc {
	return func(c *gin.Context) {
		logger.Info("[API] /user/register called")
		var req struct {
			Email    string `json:"email" binding:"required"`
			Password string `json:"password" binding:"required"`
			Code     string `json:"code" binding:"required"`
		}
		if err := c.ShouldBindJSON(&req); err != nil {
			middleware.ValidationError(c, "邮箱、密码和验证码不能为空")
			return
		}
		if !util.IsValidEmail(req.Email) {
			middleware.ValidationError(c, "邮箱格式不正确")
			return
		}
		if !util.IsStrongPassword(req.Password) {
			middleware.ValidationError(c, "密码需8位以上，含大小写字母和数字")
			return
		}
		if len(req.Code) != 6 {
			middleware.ValidationError(c, "验证码格式不正确")
			return
		}
		err := service.ActivateUserWithPassword(db, req.Email, req.Password, req.Code)
		if err != nil {
			if err == sql.ErrNoRows {
				middleware.BadRequest(c, "验证码无效或已过期")
			} else {
				middleware.InternalServerError(c, "注册失败")
			}
			return
		}
		middleware.SuccessWithMessage(c, "注册成功，请登录", nil)
	}
}

// Login godoc
// @Summary      用户登录
// @Description  用户登录接口
// @Tags         用户
// @Accept       json
// @Produce      json
// @Param        login body object{email=string,password=string} true "登录信息"
// @Success      200    {object}  model.Response
// @Router       /user/login [post]
func Login(db *sql.DB) gin.HandlerFunc {
	return func(c *gin.Context) {
		logger.Info("[API] /user/login called")
		var req struct {
			Email    string `json:"email" binding:"required"`
			Password string `json:"password" binding:"required"`
		}
		if err := c.ShouldBindJSON(&req); err != nil {
			middleware.ValidationError(c, "邮箱和密码不能为空")
			return
		}
		if !util.IsValidEmail(req.Email) {
			middleware.ValidationError(c, "邮箱格式不正确")
			return
		}
		token, err := service.LoginUser(db, req.Email, req.Password)
		if err != nil {
			middleware.Unauthorized(c, "邮箱或密码错误")
			return
		}
		if token == "not_activated" {
			middleware.Unauthorized(c, "账户未激活，请先验证邮箱")
			return
		}
		if token == "locked" {
			middleware.TooManyRequests(c, "账户暂时锁定，请稍后再试")
			return
		}
		middleware.Success(c, gin.H{"token": token})
	}
}

// ForgotPasswordSendCode godoc
// @Summary      忘记密码-发送验证码
// @Description  已注册用户发送重置密码验证码
// @Tags         用户
// @Accept       json
// @Produce      json
// @Param        email  body  object{email=string}  true  "邮箱"
// @Success      200    {object}  model.Response
// @Router       /user/forgot-password/send-code [post]
func ForgotPasswordSendCode(db *sql.DB) gin.HandlerFunc {
	return func(c *gin.Context) {
		logger.Info("[API] /user/forgot-password/send-code called")
		var req struct {
			Email string `json:"email" binding:"required"`
		}
		if err := c.ShouldBindJSON(&req); err != nil {
			middleware.ValidationError(c, "邮箱不能为空")
			return
		}
		if !util.IsValidEmail(req.Email) {
			middleware.ValidationError(c, "邮箱格式不正确")
			return
		}
		code, err := service.ForgotPasswordSendCode(db, req.Email)
		if err != nil {
			if err.Error() == "用户未注册" {
				middleware.BadRequest(c, "用户未注册")
			} else if err.Error() == "账户未激活" {
				middleware.BadRequest(c, "账户未激活")
			} else {
				middleware.InternalServerError(c, "发送验证码失败")
			}
			return
		}
		middleware.SuccessWithMessage(c, "验证码已发送", gin.H{
			"code": code,
		})
	}
}

// ForgotPasswordReset godoc
// @Summary      忘记密码-重置密码
// @Description  校验验证码并重置密码
// @Tags         用户
// @Accept       json
// @Produce      json
// @Param        reset body object{email=string,code=string,new_password=string} true "重置信息"
// @Success      200    {object}  model.Response
// @Router       /user/forgot-password/reset [post]
func ForgotPasswordReset(db *sql.DB) gin.HandlerFunc {
	return func(c *gin.Context) {
		logger.Info("[API] /user/forgot-password/reset called")
		var req struct {
			Email       string `json:"email" binding:"required"`
			Code        string `json:"code" binding:"required"`
			NewPassword string `json:"new_password" binding:"required"`
		}
		if err := c.ShouldBindJSON(&req); err != nil {
			middleware.ValidationError(c, "邮箱、验证码和新密码不能为空")
			return
		}
		if !util.IsValidEmail(req.Email) {
			middleware.ValidationError(c, "邮箱格式不正确")
			return
		}
		if !util.IsStrongPassword(req.NewPassword) {
			middleware.ValidationError(c, "新密码需8位以上，含大小写字母和数字")
			return
		}
		if len(req.Code) != 6 {
			middleware.ValidationError(c, "验证码格式不正确")
			return
		}
		err := service.ForgotPasswordReset(db, req.Email, req.Code, req.NewPassword)
		if err != nil {
			if err.Error() == "验证码无效或已过期" {
				middleware.BadRequest(c, "验证码无效或已过期")
			} else {
				middleware.InternalServerError(c, "重置密码失败")
			}
			return
		}
		middleware.SuccessWithMessage(c, "密码重置成功，请登录", nil)
	}
}