# backend-api/internal/model/response.go
package model

// Response 统一响应结构
// swagger:model
type Response struct {
	Code    int         `json:"code" example:"200"`
	Message string      `json:"message" example:"success"`
	Data    interface{} `json:"data,omitempty"`
}